All done but...
I was unable to add the comments as when I tried to add the Sneak2&3 my computer had a nervous breakdown because of the errors. but other than that everything is good. and a fix for the Sneak2&3 errors would be to delete them but I cant open the file without my computer crashing.